Set line width and label font size proportional to image size
The line width of the bounding box will be set to 1/150th of the image width or height, whichever is smaller. If this value is smaller than the previous default value of 2, 2 will be chosen.

Same for the label font size. This will be set to 1/40th of the image width or height, whichever is smaller. If this value is smaller than the previous default value of 15, 15 will be chosen.